pyproject.toml
[project]
name = "sot-prompt-experiment"
version = "0.1.0"
description = "Schema-of-Thought Prompting experiment comparing CoT methods"
requires-python = ">=3.11"
dependencies = [
    "hydra-core>=1.3.0",
    "wandb>=0.16.0",
    "openai>=1.0.0",
    "datasets>=2.14.0",
    "scikit-learn>=1.3.0",
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "matplotlib>=3.7.0",
    "seaborn>=0.12.0",
    "tqdm>=4.65.0",
    "pyyaml>=6.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
